pairing refactors
